Change hdr to header in single_picture_hdr_flag
Rename single_picture_hdr_flag to single_picture_header_flag. The hdr
abbreviation could be misinterpreted as High Dynamic Range (HDR), so it
is best to avoid it.

Also change the remaining instances of reduced_still_picture_header to
single_picture_header_flag.

Note that full_still_picture_hdr is left unchanged because it is an
encoder option in the public API and also exposed as the
--full-still-picture-hdr command-line option.

Fixes https://gitlab.com/AOMediaCodec/avm/-/issues/1020.