Remove reset_tx_size function
This was also removed in libaom: https://aomedia-review.googlesource.com/c/aom/+/167944

It was only relevant for recursive transform partitioning. As AVM uses single-level transform partitioning, it is not relevant for AVM.

This function was called only for `--aq-mode=2`  and due to obsolete logic, resulted in buffer overflow while writing. So, removing this function fixes #1081.

Also, a relevant helper function and macros are also removed.