Code cleanup for intra prediction
Fix issue #1077 

For 90 degrees and 180 degrees intra prediction, the top-left samples are acutally not used. However, the variable of top-left-sample is set to 1 for these two intra prediction angle at first, but not used later. This MR is trying to set it to the correct value for these two angles.