fix row multithreading decoder when BRU is enabled
In #1072 , it is reported that decoding failure when BRU is enabled and multi-threading decoding is also enabled. This issue is due to tile_active_mode decoded in the tile group header do not correctly passed to the individual threads. Thus all the tiles are treated as active tile ( even if they are not). This merge request fix the issue.