Fix AVX/C mismatch bug for v256_wideshuffle_8
The bug was not in SIMD. In the reference C implementation array indexing had a ternary operator as a clever shortcut.
This operation confused GCC compiler which produced different results compared to SIMD that was correct, clang also matches SIMD correctly.
Fix explicitly defines the index as int outside the array.

This caused the reported bug in https://gitlab.com/AOMediaCodec/avm/-/issues/95.