Optimize ccso_pre_compute_class_err()
In the parent code, ccso_pre_compute_class_err() was recomputed for the
same input data set to split them in different bands.

The ccso extended luma sample values are divided in number of bands and for
each band we compute the difference between the original and reconstructed
input samples.

In the refactored code, the data is pre computed for the maximum band number
and later reuse the precomputed data to compute the lower bands by combining 2
consecutive bands.

This is a bit-exact optimization, which  is giving 1.26% instruction count
reduction for encoder on speed 1 (CTC v8.0).