Reset frame number for only the CLK
1. get_true_pyr_level() uses is_key_frame instead of display order hint : this is to resolve issue #1110 (the performance difference)
2. frame_number is set 0 only for the CLK : this is resolve issues that can create different order hint than the anchor (the order hint difference)

This version is checked with "*ForwardKeyTestLarge*" and  --kf-min-dist=32 --kf-max-dist=32  --enable-fwd-kf=1 for the open GOP structure.
This version is checked for serial encoding and parallel encoding with an RA condition. 

Note that STATS_CHANGED is added just because the order hint of key frames for the AI condition is changed. The performance remains the same.
Fixes issue #1110