Fix out-of-bound access of tpl_mvs
In TIP-ref mode, in the function get_tip_block_width_with_same_mv()
the MFMVs of adjacent 16x16 blocks are checked to increase the block
size to 32x16, 64x16. However, this function did not clip the offsets
to within frame. This patch limits the row and col offsets to
'tpl_mvs' within frame.