Fix Test Data Check multi-line command break
`grep` makes the CI exit with code 1 on empty input. Remove the `grep` step and change the regex to NOT match on starting `$`, to skip matches such as

```
${LIBAVM_TEST_DATA_PATH}/niklas_1280_720_30.y4m
${AVM_TEST_OUTPUT_DIR}/twopass_encoder_${codec}.ivf
```

which cannot be part of `test-data.sha1` or `test_data_util.cmake`.

Also add `.gitlab-ci.yml` to the `changes` condition in `.gitlab-ci.yml` so that the `Test Data Check` is run for this MR.