Bugfix for setting obu_type when show_existing_frame=1 and derive_order_hint=0
When the obu_type is determined in av1_pack_bitstream() at the encoder, the obu_type can be changed even when show_existing_frame=1 due to the order of checking parameters. 

In this MR, show_existing_frame is checked at the end to make the obu_type OBU_LEADING/REGULAR_SEF when show_existing_frame=1. (show_existing_frame has the priority compared to bru, tip, or bridge)

This is an encoder change but crucial to run the decoder correctly.

Also fixes #1134