update initializing the quantizer matrices and using qm_obu only when user_defined_qm=1
Fixes #1087 
When the sequence header is activated, the quantizer matrices are initialized to be the same as the predefined matrices. 
If the quantizer matrices are signalled with the CLK(or OLK) in the same temporal unit, 
the quantizer matrices remain as they are not be reset to the predefined matrices. 

Fixes #1136 
QM_OBU is signalled only when user_defined_qm=1 not when enable_qm=1