still_picture_test.cc: Also test the AOM_VBR mode
still_picture_test.cc was testing only the AOM_Q mode. Also test the
AOM_VBR mode, which would have caught issue
https://gitlab.com/AOMediaCodec/avm/-/issues/1168. The
end_usage_checked_ member was unused. Rename it end_usage_ and use it
for this purpose.

Change the data member still_pic_info_ to be a local variable in
HandleDecodeResult() because it is only used in that method.