Nightly CI: when test_libavm fails, exit with the specific exit code
- Avoids Gitlab CI always returning the generic failure exit code 1, and
- Also avoids running any further commands on failure.

Note that we only do this for:

(a) unit tests ran with default build, used for coverage:  This includes further steps we should avoid running on failure.

(b) unit tests run with sanitizer: This also includes further steps, as well as most often runs into OOM issues described in #1215 

First step for #1215 to figure out the specific exit code for the OOM issue.